TheSeed Concept - Bing Copilot Reports
To be taken with a grain of salt. This is a concept thought up while medicating for my chronic illnesses. My elevated state came up with a crazy thought: if we're living in a simulation, then every particle in existence is an entity with data. We understand some of this with current Quantum Mechanics and such, but I think it goes deeper. I think that if you can figure out the addressing system, we can make our own simulation.

Signed-off-by: Bellok <jerioke@live.com>